Verify authentication on all applicable messages. Adds and renames models to match Proverif report
The ServerSigKey on ServerAuth and FullAuth are a bit of a hack. Verifpal doesn't
support multiple assignment, aliasing or anything that would allow a principal to
send the same information twice. That's what ServerSigKey is sending the public key
that we *already* got via pre-auth. By hashing the sig in M1's ServerSigKey, we can
model the function.

Could be revisited if Verifpal gains the ability in the future.